Add files via upload
This is the chromebook_m_ralt keyboard config. The default behavior from GalliumOS is to have the F-keys act as media keys without the overlay buttons, and take the overlay (R-alt) to make them F-keys. I added the line to remap the LWIN (Search key) to the Escape. 

I might go back and change this again to use the one that makes F-keys be F-keys by default and take the overlay key to make them media keys. Not sure. All I really wanted in life was for the "caps lock"/Search key to be an Escape key.